+++ vnext/Microsoft.ReactNative/Fabric/Composition/ParagraphComponentView.cpp
@@ -17,7 +17,6 @@
 #include <winrt/Microsoft.ReactNative.Composition.h>
 #include <winrt/Microsoft.UI.Input.h>
 #include <winrt/Windows.ApplicationModel.DataTransfer.h>
-#include "CompositionDynamicAutomationProvider.h"
 #include "CompositionHelpers.h"
 #include "RootComponentView.h"
 #include "TextDrawing.h"