+++ vnext/Microsoft.ReactNative/Fabric/Composition/ScrollViewComponentView.cpp
@@ -19,7 +19,6 @@
 #include <AutoDraw.h>
 #include <Fabric/DWriteHelpers.h>
 #include <unicode.h>
-#include "CompositionDynamicAutomationProvider.h"
 #include "JSValueReader.h"
 #include "RootComponentView.h"